patchwork_parser: move into utils.py
Get rid of the patchwork parser file, and move the parse_patch
function into utils.py
